More arrays implemented
Implemented all the necessary arrays except for the Odd Genre Out's
objective, which has a dependency on the Odd Genre out, and so will be a
bit different. Don't think there's an efficient way to do this except to
say, "if it's this genre, choose from this list" and etc. Will need to
add more content once the rest is done.

TO DO
-- implement the individual objectives related to odd genre out
-- write proper rules out
-- make more content for each category
-- polish it all